Changed USA SVG Map Logic
Moved the pertinent parameters to the Applications controller to store in a list, and allow a fully dynamic SVG to be generated in any web page. Mostly useful for the Pres, Gov and Senate races, where I want the states to change based off of who is leading at the time the user clicks on the site.