General Update June Return
Added a lot of visual changes to remove some visual bugs on the site where text was being hidden underneath the nav bar.

Started work on a new House Race functionality to where house races are depicted on their own.

Changed around administrator view of the pages so its easier to navigate and also manage races.